Convert ExportProjectButton to functional component ExportButton; Add download-ready text when export done.
